Bug #7265: Fix MGA DRI locking.
The server expects the driver to take the DRI lock in the wakeup
handler, and release it in the block handler.
The old locking scheme didn't work like this, which caused problems.
We now no longer override wakeup/block handler and just use the ones
provided by the server. This means the DRI lock will be taken
"all the time" and not just when it's needed by the driver.
MGADRISwapContext* are activated again, so the wakeup handler could be
removed.
I didn't notice any DRI performance drop, but I didn't do serious
benchmarking either.