Convert tests of pMga->Chipset to tests of single bit flags.
Several places in the driver contain if-statements that test Chipset against
several different values to determine what functionality to use.  Some of
these tests, HALCHIPSETS in particular, test as many as 6 different values.
This bloats the code size unnecessarilly.  Instead we cache the results of
some of those tests in single bit flags stored in pMga.

These changes are similar to some changes in the pci-rework branch.