xf86-video-mga: set the pan_ctl register
On my Matrox G550 most videomodes in Xorg didn't work. I found out that it
works if Xorg pixel clock is similar to the pixel clock set on framebuffer
console.

Further analysis showed that the Linux framebuffer driver sets the pan_ctl
register (the register 0xa2) according to the pixel clock, the Xorg driver
doesn't set it.

I copied the code to set the pan_ctl register from the Linux kernel to the
Xorg driver, and most videomodes in Xorg work.

The pan_ctl register is required for both analog and digital output.

The pan_ctl register is saved and restored, this is required so that we
restore text-mode screen or Linux framebuffer correctly.

Signed-off-by: Dave Airlie <airlied@redhat.com>